@@ -4,14 +4,14 @@ version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ^3.0.0
+  flutter: ^3.0.0
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flame: ^1.14.0
+  flame: ^1.15.0
   logger: ^2.0.2+1
 
   hetu_script:
