@@ -6,19 +6,20 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.3 <3.0.0"
+  sdk: ^3.0.0
+  flutter: ^3.0.0
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
-  fast_noise: ^1.0.1
+  fast_noise: ^2.0.0
 
   samsara:
     path: ../../samsara_engine/
 
-  flame: ^1.14.0
+  flame: ^1.15.0
   flutter_platform_alert: ^0.5.1
   json5: ^0.8.0
   vector_math: ^2.1.4
@@ -28,7 +29,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
 flutter:
   uses-material-design: true
