@@ -1,4 +1,4 @@
 export 'playing_card.dart';
 export 'zones/drawing_zone.dart';
 export 'zones/piled_zone.dart';
-export 'event/card_event.dart';
+export '../event/cardgame.dart';
